c / p

+++ src/main/java/com/example/demo/DTO/memberDTO.java
@@ -59,5 +59,5 @@ public class memberDTO {
 	
 	private String memberExtraAddress;
 	
-	private int memberPoint;
+	private long memberPoint;
 }
